Added cmake for ImGui and more options for build.sh
