+++ src/client/gui/imgui_impl_opengl2.cpp
@@ -35,8 +35,8 @@
 //  2016-09-10: OpenGL: Uploading font texture as RGBA32 to increase compatibility with users shaders (not ideal).
 //  2016-09-05: OpenGL: Fixed save and restore of current scissor rectangle.
 
-#include "../include/imgui.h"
-#include "../include/imgui_impl_opengl2.h"
+#include "imgui.h"
+#include "imgui_impl_opengl2.h"
 #if defined(_MSC_VER) && _MSC_VER <= 1500 // MSVC 2008 or earlier
 #include <stddef.h>     // intptr_t
 #else
