+++ src/client/gui/imgui_impl_sdl.cpp
@@ -60,8 +60,8 @@
 //  2017-08-25: Inputs: MousePos set to -FLT_MAX,-FLT_MAX when mouse is unavailable/missing (instead of -1,-1).
 //  2016-10-15: Misc: Added a void* user_data parameter to Clipboard function handlers.
 
-#include "../include/imgui.h"
-#include "../include/imgui_impl_sdl.h"
+#include "imgui.h"
+#include "imgui_impl_sdl.h"
 
 // SDL
 #include <SDL.h>
