+++ src/client/gui/my_gui.cpp
@@ -1,6 +1,6 @@
-#include "../include/imgui.h"
-#include "../include/imgui_impl_sdl.h"
-#include "../include/imgui_impl_opengl2.h"
+#include "imgui.h"
+#include "imgui_impl_sdl.h"
+#include "imgui_impl_opengl2.h"
 
 #include "SDL.h"
 #include "SDL_opengl.h"
@@ -8,7 +8,7 @@
 #include <stdio.h>
 #include <iostream>
 
-#include "../include/my_gui.hpp"
+#include "my_gui.hpp"
 
 
 void SkypeGui::ImGuiInit(){
